+++ Cookie.API/Protocol/Network/Messages/Game/Character/Choice/CharactersListRequestMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Character.Choice
 {
     public class CharactersListRequestMessage : NetworkMessage
     {
-        public const uint ProtocolId = 150;
+        public const ushort ProtocolId = 150;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
